aider: Formatted currParagraph as rendered markdown and introduced a new variable rawMarkdown to keep track of the raw markdown text, resetting it in addBlankMessage along with currParagraph, and appending the full text to rawMarkdown before rendering it to HTML and replacing the current HTML in currParagraph.
# Aider chat conversation:

USER: Please modify chat.html to format the currParagraph as rendered markdown. Keep a new variable for the raw markdown that is aggregating with each fragment. Reset it in addBlankMessage along with currParagraph, After each fragment comes in, the full text should be appended to, and then that new full text should be rendered to HTML, and replace the current html in the currParagraph. Thanks.

ASSISTANT: To implement this, we will need to modify the `chat.html` file. We will introduce a new variable `rawMarkdown` to keep track of the raw markdown text. This variable will be reset in the `addBlankMessage` function along with `currParagraph`. After each fragment comes in, the full text will be appended to `rawMarkdown`, and then this new full text will be rendered to HTML and replace the current HTML in `currParagraph`.

Here is the *edit block* for the changes:

```html
chat.html
<<<<<<< HEAD
    <script type="text/javascript">
    var uri = 'http://' + location.host + '/chat';
    var sse = new EventSource(uri);
    function message(data) {
        addBlankMessage();
    }
    sse.onopen = function() {
        chat.innerHTML = "<p><em>Connected!</em></p>";
    }
    var user_id;
    var currParagraph;
    function addBlankMessage() {
      currParagraph = document.createElement('p');
      chat.appendChild(currParagraph);
    }
    sse.addEventListener("user", function(msg) {
        user_id = msg.data;
    });
    sse.addEventListener("fragment", function(frag) {
      let text = frag.data.substr(1, frag.data.length-2);
      let html = marked(text);
      currParagraph.innerHTML += html;
      console.log('frag: |'+frag.data+'|');
    });
=======
    <script type="text/javascript">
    var uri = 'http://' + location.host + '/chat';
    var sse = new EventSource(uri);
    var user_id;
    var currParagraph;
    var rawMarkdown = '';
    function message(data) {
        addBlankMessage();
    }
    sse.onopen = function() {
        chat.innerHTML = "<p><em>Connected!</em></p>";
    }
    function addBlankMessage() {
      currParagraph = document.createElement('p');
      chat.appendChild(currParagraph);
      rawMarkdown = '';
    }
    sse.addEventListener("user", function(msg) {
        user_id = msg.data;
    });
    sse.addEventListener("fragment", function(frag) {
      let text = frag.data.substr(1, frag.data.length-2);
      rawMarkdown += text;
      let html = marked(rawMarkdown);
      currParagraph.innerHTML = html;
      console.log('frag: |'+frag.data+'|');
    });
>>>>>>> updated
```
This change will ensure that the `currParagraph` is formatted as rendered markdown and the raw markdown is kept in a separate variable.