aider: Added `#[derive(Serialize, Deserialize)]` attribute to `ChatMessage` struct.
# Aider chat conversation:

USER: Please ensure that `Serialize` and `Deserialize`

ASSISTANT:
^C KeyboardInterrupt

USER: Please ensure that `Serialize` and `Deserialize` are fully implemented for `ChatMessage`.

ASSISTANT: To ensure that `Serialize` and `Deserialize` are fully implemented for `ChatMessage`, we need to add the `#[derive(Serialize, Deserialize)]` attribute to the `ChatMessage` struct.

Here is the change:

```rust
src/chatlog/mod.rs
<<<<<<< HEAD
pub struct ChatMessage {
    pub message: ChatCompletionRequestMessage,
    pub length: usize,
}
=======
#[derive(Serialize, Deserialize)]
pub struct ChatMessage {
    pub message: ChatCompletionRequestMessage,
    pub length: usize,
}
>>>>>>> updated
```

This change will allow instances of `ChatMessage` to be serialized and deserialized using the `serde` library.