aider: Add anonymous login ability to /login and create a JavaScript file that generates a random unique anon username.
# Aider chat conversation:

USER: Please add an anonymous login ability to /login which will work simply as follows: any username starting with "anon" will ignore the password and allow login. Please also create a javascript file that generates a random unique anon username without using any external libraries.

ASSISTANT: To implement the anonymous login feature, we need to modify the `login` route in the `src/api/mod.rs` file. We will check if the username starts with "anon" and if so, we will ignore the password check.

For the JavaScript file, we will create a new file `static/js/anon_user.js` that generates a random unique anonymous username.

Here are the changes:

```rust
src/api/mod.rs
<<<<<<< SEARCH
    let login = warp::path!("login")
        .and(warp::post())
        .and(warp::body::json())
        .and_then(|credentials: Credentials| async move {

            if credentials.username == "user" && credentials.password == "password" {
                let token = jwt_util::create_token(&credentials.username)?;
                Ok(warp::reply::json(&LoginResponse { token }))
            } else {
                Err(warp::reject::custom(InvalidCredentials))
            }
        });
=======
    let login = warp::path!("login")
        .and(warp::post())
        .and(warp::body::json())
        .and_then(|credentials: Credentials| async move {

            if credentials.username.starts_with("anon") ||
               (credentials.username == "user" && credentials.password == "password") {
                let token = jwt_util::create_token(&credentials.username)?;
                Ok(warp::reply::json(&LoginResponse { token }))
            } else {
                Err(warp::reject::custom(InvalidCredentials))
            }
        });
>>>>>>> REPLACE
```

```javascript
static/js/anon_user.js
<<<<<<< SEARCH

=======
function generateAnonUsername() {
    var username = "anon";
    var possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";

    for (var i = 0; i < 5; i++)
        username += possible.charAt(Math.floor(Math.random() * possible.length));

    return username;
}

console.log(generateAnonUsername());
>>>>>>> REPLACE
```

In the JavaScript file, we generate a random string of 5 characters from the set of uppercase and lowercase letters and digits, and append it to "anon" to create a unique anonymous username. The username is then logged to the console.